pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airflow-provider-baserow"
authors = [
    {name = "Basri", email = "h@basri.me"},
]
license = {text = "Apache License 2.0"}
description = "A baserow Apache Airflow provider package"
classifiers = [
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
]
dynamic = ["version"]
requires-python = "~=3.8"
dependencies = ["apache-airflow>=2.4"]

[project.urls]
Homepage = "https://github.com/hbasria/airflow-provider-baserow/"
Source = "https://github.com/hbasria/airflow-provider-baserow/"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.entry-points.apache_airflow_provider]
provider_info = "baserow_provider.__init__:get_provider_info"

[tool.setuptools]
include-package-data = false

[tool.setuptools.dynamic]
version = {attr = "baserow_provider.__version__"}

[tool.setuptools.packages.find]
exclude = ["*tests.*", "*tests"]
